requirements.txt
typer==0.9.0
rich==13.7.1
pytest==8.0.2
openai==1.13.3
llama-index-embeddings-openai==0.1.6

